Reduced ResourceDataProcessingTrait.php responsibilities
